Improve pathfinder result
- Generate map weights based on distance from nearest wall. This will
prioritize moving away from walls when possible.
- Fix a bug in the pathfinder where intercardinal moves were treated as the
same as cardinal moves. This shouldn't be the case because diagonal
moves have to travel sqrt(2) distance instead of 1.0.
- Track rotations in the pathfinder nodes and prioritize paths with
fewer rotations when they have the same fitness value. I'm not sure if
this even does anything currently, but it's still useful to track.
- Add debug rendering of the current path.